bp: Unify http channels and exception handling
https://github.com/elastic/elasticsearch/commit/529e704b1162a232e9714ea49d13e8597f8578ba

This only ports parts of the changes because we aren't introducing the
NioTransport parts and therefore don't need the same kind of
abstractions.

We also removed a lot of the HTTP related functionality because we do
not have pluggable Rest handlers.
