+++ server/src/main/java/org/elasticsearch/transport/TcpChannel.java
@@ -19,23 +19,20 @@
 
 package org.elasticsearch.transport;
 
-import org.elasticsearch.action.ActionFuture;
-import org.elasticsearch.action.ActionListener;
-import org.elasticsearch.action.support.PlainActionFuture;
-import org.elasticsearch.cluster.node.DiscoveryNode;
-import org.elasticsearch.common.bytes.BytesReference;
-import org.elasticsearch.common.lease.Releasable;
-import org.elasticsearch.common.lease.Releasables;
-import io.crate.common.unit.TimeValue;
-
 import java.io.IOException;
 import java.net.InetSocketAddress;
-import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
+
+import org.elasticsearch.action.ActionFuture;
+import org.elasticsearch.action.ActionListener;
+import org.elasticsearch.cluster.node.DiscoveryNode;
+import org.elasticsearch.common.bytes.BytesReference;
+import org.elasticsearch.common.network.CloseableChannel;
+
+import io.crate.common.unit.TimeValue;
 
 
 /**
@@ -43,25 +40,7 @@ import java.util.concurrent.TimeoutException;
  * abstraction used by the {@link TcpTransport} and {@link TransportService}. All tcp transport
  * implementations must return channels that adhere to the required method contracts.
  */
-public interface TcpChannel extends Releasable {
-
-    /**
-     * Closes the channel. This might be an asynchronous process. There is notguarantee that the channel
-     * will be closed when this method returns. Use the {@link #addCloseListener(ActionListener)} method
-     * to implement logic that depends on knowing when the channel is closed.
-     */
-    void close();
-
-    /**
-     * Adds a listener that will be executed when the channel is closed. If the channel is still open when
-     * this listener is added, the listener will be executed by the thread that eventually closes the
-     * channel. If the channel is already closed when the listener is added the listener will immediately be
-     * executed by the thread that is attempting to add the listener.
-     *
-     * @param listener to be executed
-     */
-    void addCloseListener(ActionListener<Void> listener);
-
+public interface TcpChannel extends CloseableChannel {
 
     /**
      * This sets the low level socket option {@link java.net.StandardSocketOptions} SO_LINGER on a channel.
@@ -71,13 +50,6 @@ public interface TcpChannel extends Releasable {
      */
     void setSoLinger(int value) throws IOException;
 
-
-    /**
-     * Indicates whether a channel is currently open
-     *
-     * @return boolean indicating if channel is open
-     */
-    boolean isOpen();
 
     /**
      * Returns the local address for this channel.
@@ -94,39 +66,6 @@ public interface TcpChannel extends Releasable {
      * @param listener to execute upon send completion
      */
     void sendMessage(BytesReference reference, ActionListener<Void> listener);
-
-    /**
-     * Closes the channel.
-     *
-     * @param channel to close
-     * @param blocking indicates if we should block on channel close
-     */
-    static <C extends TcpChannel> void closeChannel(C channel, boolean blocking) {
-        closeChannels(Collections.singletonList(channel), blocking);
-    }
-
-    /**
-     * Closes the channels.
-     *
-     * @param channels to close
-     * @param blocking indicates if we should block on channel close
-     */
-    static <C extends TcpChannel> void closeChannels(List<C> channels, boolean blocking) {
-        if (blocking) {
-            ArrayList<ActionFuture<Void>> futures = new ArrayList<>(channels.size());
-            for (final C channel : channels) {
-                if (channel.isOpen()) {
-                    PlainActionFuture<Void> closeFuture = PlainActionFuture.newFuture();
-                    channel.addCloseListener(closeFuture);
-                    channel.close();
-                    futures.add(closeFuture);
-                }
-            }
-            blockOnFutures(futures);
-        } else {
-            Releasables.close(channels);
-        }
-    }
 
     /**
      * Awaits for all of the pending connections to complete. Will throw an exception if at least one of the
@@ -167,17 +106,4 @@ public interface TcpChannel extends Releasable {
         }
     }
 
-    static void blockOnFutures(List<ActionFuture<Void>> futures) {
-        for (ActionFuture<Void> future : futures) {
-            try {
-                future.get();
-            } catch (ExecutionException e) {
-                // Ignore as we are only interested in waiting for the close process to complete. Logging
-                // close exceptions happens elsewhere.
-            } catch (InterruptedException e) {
-                Thread.currentThread().interrupt();
-                throw new IllegalStateException("Future got interrupted", e);
-            }
-        }
-    }
 }
